+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to
 
 - 💄(frontend) error alert closeable on editor #284
 - ♻️(backend) Change email content #283
+- 🛂(frontend) viewers and editors can access share modal #302
 
 ## Fixed